services/analytics/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
requests>=2.26.0
scikit-learn>=0.24.2
slack_sdk>=3.12.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0